pyproject.toml
[project]
name = "vizprompt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-genai>=1.14.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["vizprompt"]

[project.scripts]
vizprompt = "vizprompt.cli.commands:main"
